backend/requirements.txt
alembic==1.9.4
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
cffi==1.15.1
click==8.1.3
fastapi==0.92.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.2
pycparser==2.21
pydantic==1.10.5
python-dotenv==0.21.1
sniffio==1.3.0
SQLAlchemy==2.0.4
starlette==0.25.0
typing_extensions==4.5.0
uvicorn==0.20.0
